Manual transcription for 1974.
Still some work to do since the bracket size and the wrestleback
size is smaller. I need to manually modify the JSON but this is what
I was able to make in the UI.

I probably also need to modify the template generation to produce fewer
columns for this bracket.